+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 <!--next-version-placeholder-->
 
+## v0.67.2 (2021-03-23)
+
+
+## v0.67.2b267 (2021-03-23)
+### Fix
+* Require kubernetes ([`18a0455`](https://github.com/ocadotechnology/aimmo/commit/18a04556b267167e3e1c06bdd93658b99cff04e7))
+
+## v0.67.1 (2021-03-23)
+
+
+## v0.67.1b259 (2021-03-23)
+
+
+## v0.67.1b256 (2021-03-23)
+### Fix
+* Fix worksheets not changing on game server (#1487) ([`5241faf`](https://github.com/ocadotechnology/aimmo/commit/5241faf086f80e63f6290ce5c337bf8c35af35e4))
+
 ## v0.67.0b247 (2021-03-17)
 ### Feature
 * Add scan_nearby, MoveTowardsAction and A* pathfinding (#1481) ([`1e37fb3`](https://github.com/ocadotechnology/aimmo/commit/1e37fb3612f3b03c9f8601153a5d7ebc25e19757))